+++ bocadillo/__init__.py
@@ -1,4 +1,5 @@
 from .api import API
 from .static import static
+from .middleware import RoutingMiddleware
 
 __version__ = '0.3.1'